+++ ChangeLog.md
@@ -14,7 +14,7 @@
     - Back to Meteor 2.9.0
     - Let the application define its own categories (todo #3)
     - Let the application decides if it displays a 'Reject All' button (todo #6)
-    - Provide a manager link (todo #28)
+    - Provide a configurable manager link (todo #28, todo #30)
 
 ### 1.3.0