+++ ChangeLog.md
@@ -4,7 +4,7 @@
 
 ## ChangeLog
 
-### 1.3.1-rc
+### 1.4.0-rc
 
     Release date: